+++ src/amplitude_experiment/flag/flag_config_api.py
@@ -9,10 +9,11 @@ import sseclient
 
 from ..connection_pool import HTTPConnectionPool, WrapperHTTPConnection
 from ..util.updater import get_duration_with_jitter
+from ..evaluation.types import EvaluationFlag
 from ..version import __version__
 
 class FlagConfigApi:
-    def get_flag_configs(self) -> List:
+    def get_flag_configs(self) -> List[EvaluationFlag]:
         pass
 
 
@@ -23,10 +24,10 @@ class FlagConfigApiV2(FlagConfigApi):
         self.flag_config_poller_request_timeout_millis = flag_config_poller_request_timeout_millis
         self.__setup_connection_pool()
 
-    def get_flag_configs(self) -> List:
+    def get_flag_configs(self) -> List[EvaluationFlag]:
         return self._get_flag_configs()
 
-    def _get_flag_configs(self) -> List:
+    def _get_flag_configs(self) -> List[EvaluationFlag]:
         conn = self._connection_pool.acquire()
         headers = {
             'Authorization': f"Api-Key {self.deployment_key}",
@@ -40,8 +41,8 @@ class FlagConfigApiV2(FlagConfigApi):
             if response.status != 200:
                 raise Exception(
                     f"[Experiment] Get flagConfigs - received error response: ${response.status}: ${response_body}")
-            flags = json.loads(response_body)
-            return flags
+            response_json = json.loads(response_body)
+            return EvaluationFlag.schema().load(response_json, many=True)
         finally:
             self._connection_pool.release(conn)